RBAC Dependency Injected

+++ routes/customer/routes_customer.py
@@ -1,5 +1,5 @@
-from fastapi import APIRouter, HTTPException, status
-from typing import List
+from fastapi import APIRouter, HTTPException, status, Depends
+from typing import List, Annotated
 from uuid import UUID
 from validations.customer import (
     CustomerRequest,
@@ -10,27 +10,36 @@ from validations.customer import (
 from validations.order import OrderResponse
 from validations.refund import RefundResponse
 from utils.db import db_dependency
+from schemas.user import User
 from schemas.customer import Customer
 from schemas.order import Order
 from schemas.refund import Refund
+from utils.auth import require_access_level, user_dependency
 
 
 router = APIRouter(prefix="/customers")
 
 
 @router.post("/add",
-             response_model=List[CustomerResponse],
+             response_model=CustomerResponse,
              status_code=status.HTTP_201_CREATED)
-async def add_customers(customers: List[CustomerRequest], db: db_dependency):
-    """Add New Customers."""
+async def add_customers(customer: CustomerRequest, db: db_dependency,
+                        current_user: user_dependency):
+    """Add New Customer."""
     try:
-        new_customers = [Customer(**customer.model_dump())
-                         for customer in customers]
-        db.add_all(new_customers)
-        db.commit()
+        new_customer = Customer(**customer.model_dump())
+        if current_user["is_internal_user"]:
+            new_customer.user_id = None
 
-        db.refresh(new_customers)
-        return [CustomerResponse.model_validate(cust) for cust in new_customers]
+        if current_user["id"] != customer.user_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                detail="Unable to Create Customer with differing User IDs")
+
+        db.add_all(new_customer)
+        db.commit()
+        db.refresh(new_customer)
+        return CustomerResponse.model_validate(new_customer)
+
     except Exception as e:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
                             detail=f"Error Adding Customers: {str(e)}")
@@ -39,7 +48,9 @@ async def add_customers(customers: List[CustomerRequest], db: db_dependency):
 @router.get("/all",
             response_model=List[CustomerResponse],
             status_code=status.HTTP_200_OK)
-async def get_customers(db: db_dependency, limit: int = 10, offset: int = 0):
+async def get_customers(db: db_dependency,
+                        current_user: Annotated[dict, Depends(require_access_level(2))],
+                        limit: int = 10, offset: int = 0):
     """Get All Customers."""
     try:
         customers = db.query(Customer).offset(offset).limit(limit).all()
@@ -53,18 +64,28 @@ async def get_customers(db: db_dependency, limit: int = 10, offset: int = 0):
 @router.get("/get/{customer_id}",
             response_model=CustomerResponse,
             status_code=status.HTTP_200_OK)
-async def get_customer(customer_id: UUID, db: db_dependency):
+async def get_customer(customer_id: UUID, db: db_dependency,
+                       current_user: user_dependency):
     """Get a Customer by ID."""
     try:
-        customer = db.query(Customer).filter(
-            Customer.id == customer_id).first()
+        customer = (
+            db.query(Customer)
+            .filter(Customer.id == customer_id)
+            .first()
+        )
         if not customer:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                                 detail="Customer Not Found")
 
+        if current_user["id"] != customer.user_id and not current_user["is_internal_user"]:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                detail="Unable to Retrieve Customer with Differing User IDs")
+
         return CustomerResponse.model_validate(customer)
+
     except HTTPException as e:
         raise e
+
     except Exception as e:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
                             detail=f"Error Fetching Customer: {str(e)}")
@@ -73,16 +94,24 @@ async def get_customer(customer_id: UUID, db: db_dependency):
 @router.put("/mod/{customer_id}",
             response_model=CustomerResponse,
             status_code=status.HTTP_202_ACCEPTED)
-async def update_customer(customer_id: UUID, update_data: CustomerUpdateRequest, db: db_dependency):
+async def update_customer(customer_id: UUID, updated_data: CustomerUpdateRequest,
+                          db: db_dependency, current_user: user_dependency):
     """Update a Customer by ID."""
     try:
-        customer = db.query(Customer).filter(
-            Customer.id == customer_id).first()
+        customer = (
+            db.query(Customer)
+            .filter(Customer.id == customer_id)
+            .first()
+        )
         if not customer:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                                 detail="Customer Not Found")
 
-        for field, value in update_data.model_dump(exclude_unset=True).items():
+        if current_user["id"] != customer.user_id and not current_user["is_internal_user"]:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                detail="Unable to Retrieve Customer with Differing User IDs")
+
+        for field, value in updated_data.model_dump(exclude_unset=True).items():
             setattr(customer, field, value)
 
         db.commit()
@@ -91,6 +120,7 @@ async def update_customer(customer_id: UUID, update_data: CustomerUpdateRequest,
 
     except HTTPException as e:
         raise e
+
     except Exception as e:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
                             detail=f"Error Updating Customer: {str(e)}")
@@ -101,14 +131,26 @@ async def update_customer(customer_id: UUID, update_data: CustomerUpdateRequest,
             status_code=status.HTTP_202_ACCEPTED)
 async def update_customer_user_id(customer_id: UUID,
                                   new_data: CustomerUserIDUpdateRequest,
-                                  db: db_dependency):
+                                  db: db_dependency,
+                                  current_user: Annotated[dict, Depends(require_access_level(2))]):
     """Update Customer's User ID."""
     try:
-        customer = db.query(Customer).filter(
-            Customer.id == customer_id).first()
+        customer = (
+            db.query(Customer)
+            .filter(Customer.id == customer_id)
+            .first()
+        )
         if not customer:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                                 detail="Customer Not Found")
+
+        if new_data.user_id == current_user["id"]:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                detail="Cannot Update Customer's User ID to Current User's ID")
+        user = db.query(User).filter(User.id == new_data.user_id).first()
+        if not user:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                                detail="User Account Not Found")
 
         customer.user_id = new_data.user_id
         db.commit()
@@ -125,7 +167,8 @@ async def update_customer_user_id(customer_id: UUID,
 
 @router.delete("/del/{customer_id}",
                status_code=status.HTTP_204_NO_CONTENT)
-async def delete_customer(customer_id: UUID, db: db_dependency):
+async def delete_customer(customer_id: UUID, db: db_dependency,
+                          current_user: user_dependency):
     """Delete a Customer by ID."""
     try:
         customer = db.query(Customer).filter(
@@ -133,6 +176,20 @@ async def delete_customer(customer_id: UUID, db: db_dependency):
         if not customer:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                                 detail="Customer Not Found")
+
+        if current_user["id"] != customer.user_id and not current_user["is_internal_user"]:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                detail="Cannot Delete Customer Account")
+
+        if current_user["is_internal_user"]:
+            user = db.query(User).filter(User.id == current_user["id"]).first()
+            if not user:
+                raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                                    detail="Current User Account Not Verifiable")
+
+            if user.level < 3:
+                raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                    detail="Cannot Delete Customer Account, Minimum Access Level should be 3")
 
         db.delete(customer)
         db.commit()
@@ -147,7 +204,8 @@ async def delete_customer(customer_id: UUID, db: db_dependency):
 
 @router.get("/orders/{customer_id}",
             status_code=status.HTTP_200_OK)
-async def get_customer_orders(customer_id: UUID, db: db_dependency):
+async def get_customer_orders(customer_id: UUID, db: db_dependency,
+                              current_user: user_dependency):
     """Get a Customer's Orders."""
     try:
         customer = db.query(Customer).filter(
@@ -157,12 +215,17 @@ async def get_customer_orders(customer_id: UUID, db: db_dependency):
                                 detail="Customer Not Found")
 
         customer_response = CustomerResponse.model_validate(customer)
+        if current_user["id"] != customer.user_id and not current_user["is_internal_user"]:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                detail="Unable to Retrieve Customer, IDs are not matched.")
+
         orders = (
             db.query(Order)
             .filter(Order.customer_id == customer_id)
             .order_by(Order.date_placed.desc())
             .all()
         )
+
         customer_response.orders = [
             OrderResponse.model_validate(order) for order in orders]
         return customer_response
@@ -177,7 +240,8 @@ async def get_customer_orders(customer_id: UUID, db: db_dependency):
 
 @router.get("/refunds/{customer_id}",
             status_code=status.HTTP_200_OK)
-async def get_customer_refunds(customer_id: UUID, db: db_dependency):
+async def get_customer_refunds(customer_id: UUID, db: db_dependency,
+                               current_user: user_dependency):
     """Get a Customer's Refunds."""
     try:
         customer = db.query(Customer).filter(
@@ -185,6 +249,10 @@ async def get_customer_refunds(customer_id: UUID, db: db_dependency):
         if not customer:
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                                 detail="Customer Not Found")
+
+        if current_user["id"] != customer.user_id and not current_user["is_internal_user"]:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                detail="Unable to Retrieve Customer, IDs are not matched.")
 
         customer_response = CustomerResponse.model_validate(customer)
         refunds = (
